third_party/WebKit/LayoutTests/fast/dom/XMLHttpRequest-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('new XMLHttpRequest().addEventListener("foo")');
shouldThrow('new XMLHttpRequest().removeEventListener("bar")');

shouldThrow('new XMLHttpRequest().addEventListener()');
shouldThrow('new XMLHttpRequest().removeEventListener()');
</script>
</body>
</html>
